+++ scripts/sql_s26_batch.sql
@@ -1,0 +1,88 @@
+BEGIN;
+
+-- Kilwa Sultanate (id=164) - 957-1513 Swahili Indian Ocean
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(164, 'Chittick, H. N. (1974). Kilwa: An Islamic Trading City on the East African Coast, 2 vols. British Institute in Eastern Africa Memoir 5. ISBN 978-0500970072.', 'https://biea.ac.uk/publications/', 'academic'),
+(164, 'Horton, M. & Middleton, J. (2000). The Swahili: The Social Landscape of a Mercantile Society. Blackwell. ISBN 978-0631189190.', 'https://www.wiley.com/en-us/The+Swahili-p-9780631189190', 'academic'),
+(164, 'Sutton, J. E. G. (1998). Kilwa: A History of the Ancient Swahili Town with a Guide to the Monuments of Kilwa Kisiwani and Adjacent Islands. Azania, 33, 113-169. DOI:10.1080/00672709809511414.', 'https://doi.org/10.1080/00672709809511414', 'academic'),
+(164, 'Pearson, M. N. (1998). Port Cities and Intruders: The Swahili Coast, India, and Portugal in the Early Modern Era. Johns Hopkins University Press. ISBN 978-0801859090.', 'https://www.press.jhu.edu/books/title/2185/port-cities-and-intruders', 'academic'),
+(164, 'Pouwels, R. L. (1987). Horn and Crescent: Cultural Change and Traditional Islam on the East African Coast, 800-1900. Cambridge University Press. ISBN 978-0521523097.', 'https://www.cambridge.org/9780521523097', 'academic');
+
+-- Chenla / Zhenla (id=374) - 550-802 pre-Angkor Cambodia
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(374, 'Vickery, M. (1998). Society, Economics, and Politics in Pre-Angkor Cambodia: The 7th-8th Centuries. Centre for East Asian Cultural Studies / Toyo Bunko. ISBN 978-4896561104.', 'https://www.toyo-bunko.or.jp', 'academic'),
+(374, 'Cœdès, G. (1968). The Indianized States of Southeast Asia. University of Hawai''i Press. ISBN 978-0824803681.', 'https://uhpress.hawaii.edu/title/the-indianized-states-of-southeast-asia/', 'academic'),
+(374, 'Chandler, D. P. (2018). A History of Cambodia (4th ed.). Westview Press / Routledge. ISBN 978-0813343631.', 'https://www.routledge.com/9780813343631', 'academic'),
+(374, 'Higham, C. (2014). Early Mainland Southeast Asia: From First Humans to Angkor. River Books. ISBN 978-6167339443.', 'https://www.riverbooksbk.com', 'academic'),
+(374, 'Jacques, C. (1979). Funan, Zhenla: The Reality Concealed by These Chinese Views of Indochina. In: R. B. Smith & W. Watson (eds.), Early South East Asia: Essays in Archaeology, History and Historical Geography, pp. 371-389. Oxford University Press. ISBN 978-0197135891.', 'https://global.oup.com', 'academic');
+
+-- Principality of Antioch (id=472) - 1098-1268 Crusader state
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(472, 'Asbridge, T. S. (2000). The Creation of the Principality of Antioch, 1098-1130. Boydell Press. ISBN 978-0851156613.', 'https://boydellandbrewer.com/9780851156613/the-creation-of-the-principality-of-antioch-1098-1130/', 'academic'),
+(472, 'Buck, A. D. (2017). The Principality of Antioch and its Frontiers in the Twelfth Century. Boydell & Brewer. ISBN 978-1783271733.', 'https://boydellandbrewer.com/9781783271733/the-principality-of-antioch-and-its-frontiers-in-the-twelfth-century/', 'academic'),
+(472, 'Runciman, S. (1951-1954). A History of the Crusades, 3 vols. Cambridge University Press. ISBN 978-0521061612.', 'https://www.cambridge.org/9780521061612', 'academic'),
+(472, 'Richard, J. (1999). The Crusades, c.1071-c.1291. Cambridge University Press. ISBN 978-0521625661.', 'https://www.cambridge.org/9780521625661', 'academic'),
+(472, 'Holt, P. M. (1986). The Age of the Crusades: The Near East from the Eleventh Century to 1517. Longman. ISBN 978-0582493025.', 'https://www.routledge.com/9780582493025', 'academic');
+
+-- Cumans / Kipchaks (id=326) - 900-1241
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(326, 'Vásáry, I. (2005). Cumans and Tatars: Oriental Military in the Pre-Ottoman Balkans, 1185-1365. Cambridge University Press. ISBN 978-0521837569.', 'https://www.cambridge.org/9780521837569', 'academic'),
+(326, 'Sinor, D. (ed.) (1990). The Cambridge History of Early Inner Asia. Cambridge University Press. ISBN 978-0521243049.', 'https://doi.org/10.1017/CHOL9780521243049', 'academic'),
+(326, 'Spinei, V. (2009). The Romanians and the Turkic Nomads North of the Danube Delta from the Tenth to the Mid-Thirteenth Century. Brill. ISBN 978-9004175365.', 'https://brill.com/view/title/16013', 'academic'),
+(326, 'Golden, P. B. (1992). An Introduction to the History of the Turkic Peoples. Turcologica 9. Otto Harrassowitz. ISBN 978-3447032742.', 'https://www.harrassowitz-verlag.de/title_2024.ahtml', 'academic'),
+(326, 'Pritsak, O. (1982). The Polovcians and Rus. Archivum Eurasiae Medii Aevi, 2, 321-380.', 'https://www.harrassowitz-verlag.de/title_2024.ahtml', 'academic');
+
+-- Tuyuhun / 吐谷浑 (id=296) - 285-670 Qinghai Xianbei
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(296, 'Beckwith, C. I. (2009). Empires of the Silk Road: A History of Central Eurasia from the Bronze Age to the Present. Princeton University Press. ISBN 978-0691135892.', 'https://press.princeton.edu/books/paperback/9780691135892/empires-of-the-silk-road', 'academic'),
+(296, 'Beckwith, C. I. (1987). The Tibetan Empire in Central Asia. Princeton University Press. ISBN 978-0691024691.', 'https://press.princeton.edu/books/paperback/9780691024691/the-tibetan-empire-in-central-asia', 'academic'),
+(296, 'Molè, G. (1970). The T''u-yü-hun from the Northern Wei to the Time of the Five Dynasties. Istituto italiano per il Medio ed Estremo Oriente. (Serie Orientale Roma XLI)', 'https://www.worldcat.org/title/tu-yu-hun-from-the-northern-wei-to-the-time-of-the-five-dynasties/oclc/152891', 'academic'),
+(296, 'Vovin, A. (2015). Some notes on the Tuyuhun (吐谷渾) language: in the footsteps of Paul Pelliot. Journal of Sino-Western Communications, 7(2), 157-166.', 'https://www.sino-western.org', 'academic'),
+(296, 'Twitchett, D. (ed.) (1979). The Cambridge History of China, Vol. 3: Sui and T''ang China, 589-906, Part I. Cambridge University Press. ISBN 978-0521214469.', 'https://doi.org/10.1017/CHOL9780521214469', 'academic');
+
+-- Igbo-Ukwu (id=562) - 800-1000 SE Nigeria bronze metallurgy
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(562, 'Shaw, T. (1970). Igbo-Ukwu: An Account of Archaeological Discoveries in Eastern Nigeria, 2 vols. Faber & Faber / Northwestern University Press. ISBN 978-0810103030.', 'https://www.nupress.northwestern.edu/9780810103030', 'academic'),
+(562, 'Shaw, T. (1977). Unearthing Igbo-Ukwu: Archaeological Discoveries in Eastern Nigeria. Oxford University Press / Ibadan. ISBN 978-0195754872.', 'https://www.worldcat.org/title/unearthing-igbo-ukwu/oclc/4007064', 'academic'),
+(562, 'Ehret, C. (2002). The Civilizations of Africa: A History to 1800. University Press of Virginia / James Currey. ISBN 978-0813920856.', 'https://www.upress.virginia.edu/title/3068/', 'academic'),
+(562, 'Chikwendu, V. E., Craddock, P. T., Farquhar, R. M., Shaw, T. & Umeji, A. C. (1989). Nigerian Sources of Copper, Lead and Tin for the Igbo-Ukwu Bronzes. Archaeometry, 31(1), 27-36. DOI:10.1111/j.1475-4754.1989.tb01051.x.', 'https://doi.org/10.1111/j.1475-4754.1989.tb01051.x', 'academic'),
+(562, 'Apley, A. (2001). Igbo-Ukwu (ca. 9th century). In: Heilbrunn Timeline of Art History. Metropolitan Museum of Art.', 'https://www.metmuseum.org/toah/hd/igbo/hd_igbo.htm', 'academic');
+
+-- Eze Nri / Kingdom of Nri (id=553) - 900-1911 Igbo theocracy
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(553, 'Onwuejeogwu, M. A. (1981). An Igbo Civilization: Nri Kingdom and Hegemony. Ethnographica / Ethiope Publishing. ISBN 978-9781230515.', 'https://www.worldcat.org/title/igbo-civilization/oclc/9116050', 'academic'),
+(553, 'Isichei, E. (1976). A History of the Igbo People. Macmillan. ISBN 978-0333189994.', 'https://www.worldcat.org/title/history-of-the-igbo-people/oclc/2426817', 'academic'),
+(553, 'Afigbo, A. E. (1981). Ropes of Sand: Studies in Igbo History and Culture. University Press of Nigeria / Oxford UP. ISBN 978-9789780013.', 'https://www.worldcat.org/title/ropes-of-sand/oclc/8053330', 'academic'),
+(553, 'Ogot, B. A. (ed.) (1992). General History of Africa, Vol. V: Africa from the Sixteenth to the Eighteenth Century. UNESCO/Heinemann. ISBN 978-0852550950.', 'https://unesdoc.unesco.org/ark:/48223/pf0000184293', 'academic'),
+(553, 'Northrup, D. (1978). Trade Without Rulers: Pre-Colonial Economic Development in South-Eastern Nigeria. Clarendon Press / Oxford University Press. ISBN 978-0198227403.', 'https://global.oup.com/academic/product/trade-without-rulers-9780198227403', 'academic');
+
+-- Kara-Khanid / Qarakhanid (id=609) - 840-1212 Turkic Central Asia
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(609, 'Asimov, M. S. & Bosworth, C. E. (eds.) (1998). History of Civilizations of Central Asia, Vol. IV: The Age of Achievement, A.D. 750 to the End of the Fifteenth Century, Part One: The historical, social and economic setting. UNESCO Publishing. ISBN 978-9231034671.', 'https://unesdoc.unesco.org/ark:/48223/pf0000119212', 'academic'),
+(609, 'Golden, P. B. (1992). An Introduction to the History of the Turkic Peoples. Turcologica 9. Otto Harrassowitz. ISBN 978-3447032742.', 'https://www.harrassowitz-verlag.de/title_2024.ahtml', 'academic'),
+(609, 'Soucek, S. (2000). A History of Inner Asia. Cambridge University Press. ISBN 978-0521657044.', 'https://www.cambridge.org/9780521657044', 'academic'),
+(609, 'Grousset, R. (1970). The Empire of the Steppes: A History of Central Asia (transl. N. Walford). Rutgers University Press. ISBN 978-0813506272.', 'https://www.rutgersuniversitypress.org/the-empire-of-the-steppes/9780813506272', 'academic'),
+(609, 'Biran, M. (2005). The Empire of the Qara Khitai in Eurasian History: Between China and the Islamic World. Cambridge University Press. ISBN 978-0521842266.', 'https://www.cambridge.org/9780521842266', 'academic');
+
+-- Grand Duchy of Lithuania (id=653) - 1236-1795
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(653, 'Rowell, S. C. (1994). Lithuania Ascending: A Pagan Empire within East-Central Europe, 1295-1345. Cambridge University Press. ISBN 978-0521450119.', 'https://www.cambridge.org/9780521450119', 'academic'),
+(653, 'Kiaupa, Z., Kiaupiene, J. & Kuncevicius, A. (2000). The History of Lithuania Before 1795. Lithuanian Institute of History / Arlila. ISBN 978-9986810137.', 'https://www.istorija.lt', 'academic'),
+(653, 'Norkus, Z. (2018). An Unproclaimed Empire: The Grand Duchy of Lithuania From the Viewpoint of Comparative Historical Sociology of Empires. Routledge. ISBN 978-1138281547.', 'https://www.routledge.com/9781138281547', 'academic'),
+(653, 'Stone, D. (2001). The Polish-Lithuanian State, 1386-1795. University of Washington Press. ISBN 978-0295980935.', 'https://uwapress.uw.edu/book/9780295980935/', 'academic'),
+(653, 'Christiansen, E. (1997). The Northern Crusades (2nd ed.). Penguin. ISBN 978-0140266535.', 'https://www.penguin.co.uk/books/14497/the-northern-crusades/9780140266535.html', 'academic');
+
+-- Cherokee / ᏣᎳᎩ (id=218) - 1540-1907
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(218, 'Perdue, T. (1998). Cherokee Women: Gender and Culture Change, 1700-1835. University of Nebraska Press. ISBN 978-0803287600.', 'https://www.nebraskapress.unl.edu/nebraska/9780803287600/', 'academic'),
+(218, 'McLoughlin, W. G. (1986). Cherokee Renascence in the New Republic. Princeton University Press. ISBN 978-0691047188.', 'https://press.princeton.edu/books/paperback/9780691047188/cherokee-renascence-in-the-new-republic', 'academic'),
+(218, 'Hudson, C. M. (1976). The Southeastern Indians. University of Tennessee Press. ISBN 978-0870492488.', 'https://utpress.org/title/the-southeastern-indians/', 'academic'),
+(218, 'Mooney, J. (1900). Myths of the Cherokee. Bureau of American Ethnology Annual Report 19, Part 1. Smithsonian / Government Printing Office.', 'https://www.gutenberg.org/ebooks/45634', 'academic'),
+(218, 'Sturtevant, W. C. & Fogelson, R. D. (eds.) (2004). Handbook of North American Indians, Vol. 14: Southeast. Smithsonian Institution. ISBN 978-0160723001.', 'https://www.bookstore.gpo.gov', 'academic');
+
+-- Confidence score upgrades
+UPDATE geo_entities SET confidence_score = 0.85 WHERE id IN (164, 218, 296, 326, 374, 472, 553, 562, 609, 653);
+
+COMMIT;
+
+SELECT g.id, g.name_original, g.confidence_score, (SELECT count(*) FROM sources s WHERE s.entity_id=g.id) AS n_src FROM geo_entities g WHERE g.id IN (164, 218, 296, 326, 374, 472, 553, 562, 609, 653) ORDER BY g.id;